pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
requires-python = ">=3.11"
dependencies = [
    "arabic-reshaper>=3.0.0",
    "docx>=0.2.4",
    "google-generativeai>=0.8.5",
    "pdfplumber>=0.11.7",
    "python-bidi>=0.6.6",
    "python-docx>=1.2.0",
    "python-telegram-bot>=22.3",
    "reportlab>=4.4.3",
    "telegram>=0.0.1",
]
